requirements.txt
langchain
langchain_community
python_dotenv
langchain_core
streamlit
-e .
